refactor(core): extract verifier path to rag_verifier (#465)
Closes #465. Continues the rag_core.py decomposition series after
PR-H1a (#460) and PR-H1b (#462).

Moves the verifier surface from rag_core.py:1763-1965 to new
rag_verifier.py (~203 LOC):
- verify_evidence — main verifier with the partial-topic grounding
  policy (ADR 0004 strict→relaxed staging)
- verification_topics / specific_topics / metadata_terms_for_verification
- EVIDENCE_BOUNDARY constant + 3 instruction-pattern regexes
- neutralize_instruction_patterns — ADR 0008 evidence-side defense
- evidence_text_for_verification, evidence_has_topic
- PARTIAL_TOPIC_GROUNDING_MIN_FRACTION / _MIN_MATCHED / _REASON
  policy constants + context comment

Dependency handling
-------------------
Module-level imports in rag_verifier (all safe, no path back to
rag_core):
- korean_lexicon (METADATA_EVIDENCE_LABELS, METADATA_GENERIC_TOKENS,
  TOPIC_KEYWORDS, VERIFICATION_INTENT_TOKENS)
- text_normalize (expand_forms, normalize_text)

Late-imports inside specific functions (rag_core symbols that serve
many non-verifier call sites in rag_core — analysis path, metadata
matching, claim builders — and stay there as canonical homes):
- verification_topics: normalize_metadata_token, ordered_unique
- metadata_terms_for_verification: metadata_tokens

Re-exports from rag_core
------------------------
External callers keep working unchanged. Re-export list:
- EVIDENCE_BOUNDARY (consumed by tests/test_synthetic_judge.py,
  tests/test_prompt_injection_regression.py, scripts/llm_judge.py,
  eval/synthetic_judge.py)
- neutralize_instruction_patterns (scripts/llm_judge.py,
  eval/synthetic_judge.py)
- verify_evidence, specific_topics, verification_topics +
  internals + 3 policy constants (called from rag_core.py:3462
  and 2214 — orchestration + answer generation paths inside this
  module)

Cross-ref updates
-----------------
- CLAUDE.md repository map adds rag_verifier.py entry
- scripts/_governance.py LOAD_BEARING_PATHS adds rag_verifier.py
- .github/pull_request_template.md load-bearing hint adds rag_verifier.py
  (so tests/test_governance.py::test_pr_template_mentions_all_canonical_entries
  stays green)
- ADR 0004 Related field points at rag_verifier.py:verify_evidence
- ADR 0008 Related field points at rag_verifier.py for
  evidence_text_for_verification / neutralize_instruction_patterns /
  EVIDENCE_BOUNDARY; also adds ADR 0028 as complementary query-side
  defense

LOC accounting
--------------
rag_core.py: 3,819 → 3,635 (-184)
rag_verifier.py: 0 → 268 (new)
Net file +1, single-module god-file pressure decreased.

Verification
------------
Full local pytest 838 passed (1 skipped, 0 failed). All five
verifier-touching regression gates stay green:
- tests/test_naive_baseline_ranking_invariance.py (ADR 0001 golden)
- tests/test_retrieval_loop_regression.py
- tests/test_prompt_injection_regression.py (ADR 0008 surface)
- tests/test_synthetic_judge.py (EVIDENCE_BOUNDARY contract)
- tests/test_langgraph_orchestrator_regression.py
