+++ CLAUDE.md
@@ -46,6 +46,7 @@ Supporting:
 - `rag_vector_store.py` — `VectorStore` Protocol behind the embeddings sidecar (issue #232, Stage 1 of #176). `BIDMATE_INDEX_BACKEND=memory` is the only supported value today; `qdrant` / `pgvector` are reserved for follow-up PRs.
 - `rag_reranker.py` — `Reranker` Protocol + default `CrossEncoderReranker` adapter (issue #345, follow-up to #332). Wraps `rag_rerank.rerank` so future HyDE / LLM-as-reranker impls plug in without touching `rag_retrieval.apply_fusion_and_reranking`.
 - `rag_retrieval.py` — retrieval pipeline extracted from `rag_core.py` across PR-H1a (issue #459) + PR-H1b (issue #461). Owns `retrieve_candidates` (candidate generation), the four similarity primitives (`embed_query_for_index`, `dense_similarity`, `lexical_similarity`, `metadata_similarity`), BM25 surface (`bm25_scores_for_index`, `get_or_build_bm25` + `_*` helpers), and the post-retrieval stack (`apply_fusion_and_reranking`, `apply_comparison_balance`, `reassemble_parent_sections`). Late-imports `tokenize` / `DEFAULT_EMBEDDING_MODEL` / `DEFAULT_HASH_DIM` / `embed_texts` / `hashing_embeddings` / `comparison_targets_for_analysis` / `normalize_regions` / `normalize_page_span` from `rag_core` to avoid circular imports — these helpers serve many non-retrieval rag_core call sites and stay there as canonical homes.
+- `rag_verifier.py` — verifier path extracted from `rag_core.py` (issue #465, PR-J1). Owns `verify_evidence` (main verifier + partial-topic grounding policy per ADR 0004), the topic-extraction helpers (`verification_topics`, `specific_topics`, `metadata_terms_for_verification`), the `EVIDENCE_BOUNDARY` constant + 3 instruction-pattern regexes, `neutralize_instruction_patterns` (ADR 0008 evidence-side defense), `evidence_text_for_verification`, `evidence_has_topic`, and the `PARTIAL_TOPIC_GROUNDING_*` policy constants. Direct imports: `korean_lexicon` (METADATA_EVIDENCE_LABELS / METADATA_GENERIC_TOKENS / TOPIC_KEYWORDS / VERIFICATION_INTENT_TOKENS) + `text_normalize` (expand_forms / normalize_text). Late-imports `normalize_metadata_token` / `metadata_tokens` / `ordered_unique` from `rag_core` so the module stays a leaf in the dependency graph. `rag_core` re-exports every public name + 3 constants so `tests/test_synthetic_judge.py` / `tests/test_prompt_injection_regression.py` / `scripts/llm_judge.py` / `eval/synthetic_judge.py` (which import `EVIDENCE_BOUNDARY` and `neutralize_instruction_patterns` from `rag_core`) keep working unchanged.
 - `rag_query_expansion.py` — `QueryExpander` Protocol + default `IdentityExpander` + opt-in `HyDEExpander` (issue #396, ADR 0023). Plugs in before the dense-embedding call in `rag_retrieval.retrieve_candidates`; BM25 / lexical / metadata paths consume `analysis.tokens` and remain invariant. Identity default preserves the ADR 0001 `naive_baseline` bit-identical golden.
 - `scripts/` — `build_index.py`, `update_readme_metrics.py`, `run_real_eval_delta.py`, etc.
 - `data/raw/` → `data/index/` → `outputs/` → `reports/` (pipeline artifacts).
