+++ scripts/_governance.py
@@ -33,6 +33,7 @@ import sys
 LOAD_BEARING_PATHS: list[str] = [
     "rag_core.py",
     "rag_retrieval.py",
+    "rag_verifier.py",
     "ingestion.py",
     "visual_ingestion.py",
     "eval/",
